moved pokemon stats into its own component
